+++ aaem/components/biomass_pellet.py
@@ -165,13 +165,13 @@ def component_summary (coms, res_dir):
         'notes'
         ]
             
-    data = DataFrame(out,columns = cols).set_index('Community').round(2)
+    data = DataFrame(out,columns = cols).set_index('Community')#.round(2)
     f_name = os.path.join(res_dir,
                 COMPONENT_NAME.replace(" ","_") + '_summary.csv')
     #~ fd = open(f_name,'w')
     #~ fd.write(("# " + COMPONENT_NAME + " summary\n"))
     #~ fd.close()
-    data.to_csv(f_name, mode='a')
+    data.to_csv(f_name, mode='w')
 
 #   do a find and replace on ComponentName to name of component 
 # (i.e. 'ResidentialBuildings')