+++ aaem/components/solar_power.py
@@ -241,13 +241,13 @@ def component_summary (coms, res_dir):
             'notes']
     
     
-    data = DataFrame(out,columns = cols).set_index('Community').round(2)
+    data = DataFrame(out,columns = cols).set_index('Community')#.round(2)
     f_name = os.path.join(res_dir,
                 'solar_power_summary.csv')
     #~ fd = open(f_name,'w')
     #~ fd.write(("# solar summary\n"))
     #~ fd.close()
-    data.to_csv(f_name, mode='a')
+    data.to_csv(f_name, mode='w')
     
 ## list of prerequisites for module
 prereq_comps = []